Use size_t where appropriate.
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@22868 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/krb5/n-fold.c
@@ -38,8 +38,9 @@ static krb5_error_code
 rr13(unsigned char *buf, size_t len)
 {
     unsigned char *tmp;
-    int bytes = (len + 7) / 8;
-    int i;
+    size_t bytes = (len + 7) / 8;
+    size_t i;
+
     if(len == 0)
 	return 0;
     {
@@ -84,7 +85,7 @@ rr13(unsigned char *buf, size_t len)
 static void
 add1(unsigned char *a, unsigned char *b, size_t len)
 {
-    int i;
+    size_t i;
     int carry = 0;
     for(i = len - 1; i >= 0; i--){
 	int x = a[i] + b[i] + carry;
